Priority Queue AVL Tree Based

+++ header/Spare_Cube.h
@@ -6,11 +6,11 @@
 template <class T>
 struct Depth_Node
 {
-    ll depth;
+    int depth;
     Sparse_Matrix<T> matrix;
     Depth_Node<T> *next{};
     Depth_Node<T> *prev{};
-    Depth_Node(ll d, ll r, ll c) : depth(d), matrix(r, c) {}
+    Depth_Node(int d, int r, int c) : depth(d), matrix(r, c) {}
 };
 
 template <class type>
@@ -18,13 +18,13 @@ class Spare_Cube
 {
     Depth_Node<type> *head{};
     Depth_Node<type> *tail{};
-    ll length = 0;
-    ll rows{};
-    ll cols{};
-    ll depths{};
-    Depth_Node<type> *get_depth(ll depth, bool flag);
+    int length = 0;
+    int rows{};
+    int cols{};
+    int depths{};
+    Depth_Node<type> *get_depth(int depth, bool flag);
     void link(Depth_Node<type> *first, Depth_Node<type> *second);
-    Depth_Node<type> *add_node_between_node_and_next(Depth_Node<type> *node_before, ll depth);
+    Depth_Node<type> *add_node_between_node_and_next(Depth_Node<type> *node_before, int depth);
 
 public:
     Spare_Cube(int row, int col, int depth);
